Create Resume Ranker
Resume Ranker is the process of ranking a candidate for a given role based on his or her education, experience and other information obtained on his or her resume.
It is a form of pattern matching between a job’s descriptions or requirements and the qualification of a candidate based on their resume.